Seperate history module into directory
